code in logs displayed refactored, some functions moved to handlers file
