+++ packages/wheelhouse-frontend/src/index.js
@@ -4,7 +4,6 @@ import { HashRouter as Router, Route } from "react-router-dom";
 import "milligram";
 import { Provider } from "react-redux";
 import DevelopmentDataDisplay from "./components/development-data/DevelopmentDataDisplay"; 
-// import Header from "./components/reusables/Header";
 import LogsDataDisplay from "./components/logs-data/LogsDataDisplay";
 import PodsDataDisplay from "./components/kubectl-data/PodsDataDisplay"; 
 import store from "./store";
